Reference

Types

Elysivm.LSSVC — Type

LSSVC()
LSSVC(; kernel="rbf", γ=1.0, σ=1.0)

The type to hold a Least Squares Support Vector Classifier.

Fields

kernel::String: The kind of kernel to use for the non-linear mapping of the data.

γ::Float64: The gamma hyperparameter that is intrinsic of the Least Squares version of the Support Vector Machines.

σ::Float64: The hyperparameter for the RBF kernel.

Keywords

kernel: A string to denote the kernel to be used.

γ: A float value to assign the gamma hyperparameter.

σ: A float value to assign the sigma hyperparameter.

Elysivm.LSSVR — Type

LSSVR()
LSSVR(; kernel="rbf", γ=1.0, σ=1.0)

The type to hold a Least Squares Support Vector Regressor.

Fields

kernel::String: The kind of kernel to use for the non-linear mapping of the data.

γ::Float64: The gamma hyperparameter that is intrinsic of the Least Squares version of the Support Vector Machines.

σ::Float64: The hyperparameter for the RBF kernel.

Keywords

kernel: A string to denote the kernel to be used.

γ: A float value to assign the gamma hyperparameter.

σ: A float value to assign the sigma hyperparameter.

Elysivm.SVM — Type

SVM

A super type for both classifiers and regressors that are implemented as Support Vector Machines.

Methods

Elysivm.svmpredict — Method

svmpredict(svm::LSSVC, fits, xnew::AbstractMatrix) -> AbstractArray

Uses the information obtained from svmtrain such as the bias and weights to construct a decision function and predict new class values.

Arguments

svm::LSSVC: The Support Vector Machine that contains the hyperparameters, as well as the kernel to be used.

fits: It can be any container data structure but it must have four elements: x, the data matrix; y, the labels vector; α, the weights; and b, the bias.

xnew::AbstractMatrix: The data matrix that contains the new instances to be predicted.

Returns

Array: The labels corresponding to the prediction to each of the instances in xnew.

Elysivm.svmpredict — Method

svmpredict(svm::LSSVR, fits, xnew::AbstractMatrix) -> AbstractArray

Uses the information obtained from svmtrain such as the bias and weights to construct a decision function and predict the new values of the function.

Arguments

svm::LSSVR: The Support Vector Machine that contains the hyperparameters, as well as the kernel to be used.

fits: It can be any container data structure but it must have four elements: x, the data matrix; y, the labels vector; α, the weights; and b, the bias.

xnew::AbstractMatrix: The data matrix that contains the new instances to be predicted.

Returns

Array: The labels corresponding to the prediction to each of the instances in xnew.

Elysivm.svmtrain — Method

svmtrain(svm::LSSVC, x::AbstractMatrix, y::AbstractVector) -> Tuple

Solves a Least Squares Support Vector Classification problem using the Conjugate Gradient method. In particular, it uses the Lanczos process due to the fact that the matrices are symmetric.

Arguments

svm::LSSVC: The Support Vector Machine that contains the hyperparameters, as well as the kernel to be used.

x::AbstractMatrix: The data matrix with the features. It is expected that this array is already standardized, i.e. the mean for each feature is zero and its standard deviation is one.

y::AbstractVector: A vector that contains the classes. It is expected that there are only two classes, -1 and 1.

Returns

Tuple: A tuple containing x, y and the following two elements:

b: Contains the bias for the decision function.

α: Contains the weights for the decision function.

Elysivm.svmtrain — Method

svmtrain(svm::LSSVR, x::AbstractMatrix, y::AbstractVector) -> Tuple

Solves a Least Squares Support Vector Regression problem using the Conjugate Gradient method. In particular, it uses the Lanczos process due to the fact that the matrices are symmetric.

Arguments

svm::LSSVR: The Support Vector Machine that contains the hyperparameters, as well as the kernel to be used.

x::AbstractMatrix: The data matrix with the features. It is expected that this array is already standardized, i.e. the mean for each feature is zero and its standard deviation is one.

y::AbstractVector: A vector that contains the classes. It is expected that there are only two classes, -1 and 1.

Returns

Tuple: A tuple containing x and the following two elements:

b: Contains the bias for the decision function.

α: Contains the weights for the decision function.